Marked connection as outside of transaction upon commit and rollback
Failure to update the wrappers transaction status when closing a transaction means subsiquent transaction calls will be performed outside of a transaction, then error once the second transaction is committed or rolled back. This will resolve these issues by making sure the wrapper is aware of the updated transaction status once the commit is performed.